requirements.txt
flask==2.0.1
PyPDF2==3.0.1
nltk==3.8.1
google-generativeai==0.7.0
pypdf
python-docx 